Create CSSSpecificity.html
CSS specificity determines the priority of CSS rules when multiple rules apply to the same element. It is calculated based on the types of selectors used in the rule. For example:

Inline styles have the highest specificity.

ID selectors are more specific than class selectors.

Class selectors, attributes, and pseudo-classes are more specific than element selectors.